drivers: net: xilinx_emaclite: Fix -Wpointer-to-int-cast warnings with W=1
drivers/net/ethernet//xilinx/xilinx_emaclite.c:341:35: warning: cast from
pointer to integer of different size [-Wpointer-to-int-cast]
  341 |   addr = (void __iomem __force *)((u32 __force)addr ^

Use uintptr_t instead of u32 to avoid problems on 64 bit systems.

Also, cast the address to an unsigned long for printing.

Rebased on top of Xilinx tree and fixed conflicts.

Fix conflict by replacing ulong to uintptr.

State: upstream (eccd5403814b4e762e270ef0464bb86fb217b1bf)

+++ drivers/net/ethernet/xilinx/xilinx_emaclite.c
@@ -328,7 +328,7 @@ static int xemaclite_send_data(struct net_local *drvdata, u8 *data,
 		 * if it is configured in HW
 		 */
 
-		addr = (void __iomem __force *)((ulong __force)addr ^
+		addr = (void __iomem __force *)((uintptr_t __force)addr ^
 						 XEL_BUFFER_OFFSET);
 		reg_data = xemaclite_readl(addr + XEL_TSR_OFFSET);
 
@@ -390,8 +390,9 @@ static u16 xemaclite_recv_data(struct net_local *drvdata, u8 *data, int maxlen)
 		 * will correct on subsequent calls
 		 */
 		if (drvdata->rx_ping_pong != 0)
-			addr = (void __iomem __force *)((ulong __force)addr ^
-							 XEL_BUFFER_OFFSET);
+			addr = (void __iomem __force *)
+				((uintptr_t __force)addr ^
+				 XEL_BUFFER_OFFSET);
 		else
 			return 0;	/* No data was available */
